Aggiustamenti per download e rimozione eventi da DLQ

+++ local-start-docker.sh
@@ -8,7 +8,8 @@
 #     esecuzione della run indicata dal secondo parametro.
 
 if ( [ "$#" -ne "3" ]) then
-  echo "Usage $0 <env_file> <output_subfolder> <execution_or_download>"
+  echo "Usage $0 <env_file> <output_subfolder> <command>"
+  echo " where command could be: run download, clean_dlq"
   exit 1
 fi
 
@@ -18,7 +19,7 @@ credentials_files=$1
 output_subfolder=$2
 execution_or_download=$3
 
-if ( [ "$execution_or_download" != "download" ] ) then
+if ( [ "$execution_or_download" == "run" ] ) then
   mkdir -p $(pwd)/outputs/$output_subfolder
 
   docker run --rm -ti --env-file $credentials_files \
@@ -30,7 +31,7 @@ if ( [ "$execution_or_download" != "download" ] ) then
     --http-debug \
     --console-output=/outputs/console-output.txt \
     --log-output=file=/outputs/http-output.json
-else
+elif ( [ "$execution_or_download" == "download" ] ) then
   docker run --rm -ti \
     -v $(pwd)/outputs/$output_subfolder:/outputs/ \
     -v "${LOCAL_HOME}/.aws":/root/.aws \
@@ -38,9 +39,18 @@ else
     k6 \
     -f /outputs/ \
     -r eu-south-1 \
-    -p sso_pn-core-dev
-
-  #./get_test_metrics/get_test_metrics_by_folder.sh -f $(pwd)/outputs/$output_subfolder -p sso_pn-core-dev -r eu-south-1
+    -p sso_pn-core-dev \
+    -c sso_pn-confinfo-dev
+elif ( [ "$execution_or_download" == "clean_dlq" ] ) then
+  docker run --rm -ti \
+    -v $(pwd)/outputs/$output_subfolder:/outputs/ \
+    -v "${LOCAL_HOME}/.aws":/root/.aws \
+    --entrypoint /tests/get_test_metrics/clean_dlq.sh \
+    k6 \
+    -f /outputs/ \
+    -r eu-south-1 \
+    -p sso_pn-core-dev \
+    -c sso_pn-confinfo-dev
 fi
